Double buffering for GraphBuffer
To avoid having the swing thread wait for a running render operation,
we use two BufferedImages, and just swap them on completing render.
This means that Swing should never have to wait a significant amount of
time to repaint. This also means that one graph implementation should
be suitable for graphs with few or many series - no need to separate
them by whether the rendering is performed in the swing thread or
background thread - there's no reason to ever render in the swing
thread.