Implement retina support and log axes
Retina support is relatively simple - we always render to a
BufferedImage offscreen, so if we are on a retina display we can make a
double-size image and render with a transform so that everything is
scaled up by two. We then scale this back down by 0.5 when we actually
draw a component. Note that this doesn't actually require an exact 2x
scale - the trickiest bit is detecting the correct scaling for the
screen, we do this by looking at the font rendering context of the
graphics using SwingView.scaleOf. This has been tested using a MacBook
with retina screen and external non-retina screen.

Also add implementation of log scaling - works reasonably well but
needs some polish around handling of enforcing a minimum value on log
axes.