Extend subpixel rendering for high DPI screens
We now draw lines and fonts to subpixel accuracy when drawing in high
quality mode to high DPI screens. This avoids jittering of positions,
most noticeably on graph axes. This is disabled on low dpi screens
since it will look blurry, and in low quality mode, since it may be
slower.